Created currentAnswers state in RegisterForm
'currentAnswers' will keep track of the unfinished inputs given by the
user.

Once the form is submitted, it will be type-casted to the Answers interface
which has no optional fields.

With this commit, we don't need to initialize a populated answer object
state like the previous version of this file.